Track propagation time using particle velocity
Candidate propagation time was previously calculated assuming that every particle travels at the speed of light. This approximation is valid for ultrarelativistic particles but produces incorrect flight times in the non-relativistic regime.

Elapsed propagation time is now calculated using the velocity of the current ParticleState. The creation time of secondaries produced inside a propagation step is reconstructed using the parent velocity, with explicit handling of zero-velocity particles. Candidate time is stored as long double to reduce precision loss during long simulations.

+++ include/crpropa/Candidate.h
@@ -47,7 +47,7 @@ private:
 	double currentStep; /**< Size of the currently performed step in [m] comoving units */
 	double nextStep; /**< Proposed size of the next propagation step in [m] comoving units */
 	std::string tagOrigin; /**< Name of interaction/source process which created this candidate*/
-	double time; /**< Time [s] that has passed in the laboratory frame of reference */
+	long double time; /**< Time [s] that has passed in the laboratory frame of reference */
 
 	static uint64_t nextSerialNumber;
 	uint64_t serialNumber;
@@ -150,12 +150,12 @@ public:
 	 since CRPropa assumes lightspeed in every case both TrajectoryLength and Time are equal.
 	 @param t  Time in seconds
 	 */
-	void setTime(double t);
+	void setTime(long double t);
 	/** Returns the time of the candidate.
 	 * The time is tracked alongside TrajectoryLength by dividing the current TrajecoryLength by c
 	 * @return Current time in seconds
 	 */
-	double getTime() const;
+	long double getTime() const;
 
 	/**
 	 Make a bid for the next step size: the lowest wins.
